+++ spec/onlyoffice_webdriver_wrapper/web_driver/chrome_version_helper_spec.rb
@@ -3,7 +3,7 @@
 require 'spec_helper'
 
 describe OnlyofficeWebdriverWrapper::WebDriver, '#chrome_version_helper' do
-  let(:webdriver) { described_class.new(:chrome) }
+  let(:webdriver) { described_class.new(:chrome, do_not_start_browser: true) }
 
   it 'Chrome version is not zero' do
     expect(webdriver.chrome_version).not_to eq(webdriver.unknown_chrome_version)