+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### New Features
 
 * Add ability to get chrome version
+* Add ability to run `chromedriver` of different version
+* Add option to `WebDriver.new` to not start option
 
 ### Changes